Use atom keymap watcher and refactor
Previously, the code watched every save and checked if it was valid. I
didn't test, but I assumed this performed worse than atom's build-in
keymap watch system. Also, the code was somewhat inelegant.

Besides that, I also restructured the code to be slightly more modular.
Previously, I had to define functions before they were used, which gave
me less freedom in the code structure. I fixed this by placing all the
functions in the global namespace such that they will all be loaded
before the export. While this does clutter the global namespace, it's
acceptable for a small program.